moving js types in a new folder + small fixes
