+++ .gitignore
@@ -1,1 +1,3 @@
 *.code-workspace
+*.log
+test*
