+++ javascript/.gitignore
@@ -1,3 +1,2 @@
-test*
 node_modules/
 dist/
